requirements.txt
Django==1.7
django-phonenumber-field==0.6
django-twilio==0.7.1
